# openapi/toeslagen.yaml
openapi: 3.1.0
info:
  title: Toeslagen Rule API
  description: Auditbare juridische rule service voor toeslagen
  version: 1.0.0
paths:
  /v1/health:
    get:
      summary: Health
      operationId: health_v1_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /v1/toeslagen/versions:
    get:
      summary: Get Versions
      operationId: get_versions_v1_toeslagen_versions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /v1/rules/{rule_id}:
    get:
      summary: Get Rule
      operationId: get_rule_v1_rules__rule_id__get
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          title: Rule Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/audit/{calculation_id}:
    get:
      summary: Get Audit
      operationId: get_audit_v1_audit__calculation_id__get
      parameters:
      - name: calculation_id
        in: path
        required: true
        schema:
          type: string
          title: Calculation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/toeslagen/calculate:
    post:
      summary: Calculate Toeslag
      operationId: calculate_toeslag_v1_toeslagen_calculate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToeslagRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/toeslagen/explain:
    post:
      summary: Explain Toeslag
      operationId: explain_toeslag_v1_toeslagen_explain_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToeslagRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ToeslagRequest:
      properties:
        toeslagType:
          type: string
          title: Toeslagtype
          description: zorgtoeslag, huurtoeslag, kinderopvangtoeslag, kindgebonden_budget
        leeftijd:
          anyOf:
          - type: integer
          - type: 'null'
          title: Leeftijd
        alleenstaande:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Alleenstaande
        toeslagenpartner:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Toeslagenpartner
        inkomen:
          anyOf:
          - type: number
          - type: 'null'
          title: Inkomen
        huur:
          anyOf:
          - type: number
          - type: 'null'
          title: Huur
        kindOpOpvang:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Kindopopvang
        heeftKind:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Heeftkind
        nationaliteit:
          anyOf:
          - type: string
          - type: 'null'
          title: Nationaliteit
        verblijfsrecht:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Verblijfsrecht
        gedetineerd:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Gedetineerd
        woonsituatie:
          anyOf:
          - type: string
          - type: 'null'
          title: Woonsituatie
      type: object
      required:
      - toeslagType
      title: ToeslagRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError